Moved definition_manager.py to the api submodule
