+++ amulet/api/definition_manager.py
