+++ amulet/version_definitions/__init__.py
@@ -1,1 +1,1 @@
-from .definition_manager import DefinitionManager
+
